+++ EasyDotnet.Application/Interfaces/ISolutionService.cs
@@ -1,0 +1,8 @@
+using EasyDotnet.Domain.Models.Solution;
+
+namespace EasyDotnet.Application.Interfaces;
+
+public interface ISolutionService
+{
+  List<SolutionFileProject> GetProjectsFromSolutionFile(string solutionFilePath);
+}
