+++ EasyDotnet.Domain/Models/Solution/SolutionFileProject.cs
@@ -1,0 +1,6 @@
+namespace EasyDotnet.Domain.Models.Solution;
+
+public sealed record SolutionFileProject(
+    string ProjectName,
+    string AbsolutePath
+);
